RAC: remove old calendar files
