+++ packages/components/src/Calendar/Calendar.tsx
@@ -1,149 +1,91 @@
-import { createCalendar } from '@internationalized/date';
-import { useRef, useState } from 'react';
-
+import React, { useState } from 'react';
+import type RAC from 'react-aria-components';
 import {
-  AriaCalendarProps,
-  DateValue,
-  useCalendar,
-} from '@react-aria/calendar';
-import { useDateFormatter, useLocale } from '@react-aria/i18n';
+  Button,
+  Calendar,
+  CalendarCell,
+  CalendarGrid,
+  CalendarGridBody,
+} from 'react-aria-components';
 
-import { useCalendarState } from '@react-stately/calendar';
+import { DateValue } from '@react-aria/calendar';
 
-import { ChevronDown, ChevronLeft, ChevronRight } from '@marigold/icons';
-import { cn, useClassNames, useStateProps } from '@marigold/system';
+import { ChevronLeft, ChevronRight } from '@marigold/icons';
+import { WidthProp, cn, useClassNames } from '@marigold/system';
 
-import { Button } from '../Button';
-import { CalendarGrid } from './CalendarGrid';
+import { CalendarButtonListBox } from './CalendarButtonListBox';
+import { CalendarGridHeader } from './CalendarGridHeader';
 import MonthDropdown from './MonthDropdown';
 import YearDropdown from './YearDropdown';
 
 // Props
 // ---------------
+type RemovedProps = 'isDisabled' | 'isReadOnly' | 'isInvalid';
 export interface CalendarProps
-  extends Omit<AriaCalendarProps<DateValue>, 'isDisabled' | 'isReadOnly'> {
+  extends Omit<RAC.CalendarProps<DateValue>, RemovedProps> {
   disabled?: boolean;
   readOnly?: boolean;
   variant?: string;
   size?: string;
+  width?: WidthProp['width'];
+  error?: boolean;
+  errorMessage?: string;
 }
 
-const buttonStyles = 'flex items-center justify-between gap-1 overflow-hidden';
-
+type ViewMapKeys = 'month' | 'year';
 // Component
 // ---------------
-export const Calendar = ({
+export const _Calendar = ({
   disabled,
   readOnly,
+  error,
   size,
   variant,
   ...rest
 }: CalendarProps) => {
-  const { locale } = useLocale();
-  const props: AriaCalendarProps<DateValue> = {
+  const props: RAC.CalendarProps<DateValue> = {
     isDisabled: disabled,
     isReadOnly: readOnly,
+    isInvalid: error,
     ...rest,
   };
 
-  const state = useCalendarState({
-    ...props,
-    locale,
-    createCalendar,
-  });
-  const ref = useRef(null);
-  const { calendarProps, prevButtonProps, nextButtonProps } = useCalendar(
-    props,
-    state
-  );
-  // destructure isDisabled to avoid passing it to the component and being used on dom element
-  const {
-    isDisabled: prevIsDisabled,
-    onFocusChange: prevFocusChange,
-    ...prevPropsRest
-  } = prevButtonProps;
-
-  const {
-    isDisabled: nextIsDisabled,
-    onFocusChange: nextFocusChange,
-    ...nextPropsRest
-  } = nextButtonProps;
-
-  const calendarState = useStateProps({
-    disabled: state.isDisabled,
-    focusVisible: state.isFocused,
-  });
-
   const classNames = useClassNames({ component: 'Calendar' });
-  const { select: selectClassNames } = useClassNames({ component: 'Select' });
+  const buttonClassNames = useClassNames({ component: 'Button' });
 
   const [selectedDropdown, setSelectedDropdown] = useState<
-    string | undefined
+    ViewMapKeys | undefined
   >();
 
-  let months = [];
-  let formatter = useDateFormatter({
-    month: 'long',
-    timeZone: state.timeZone,
-  });
-  let numMonths = state.focusedDate.calendar.getMonthsInYear(state.focusedDate);
-
-  for (let i = 1; i <= numMonths; i++) {
-    let date = state.focusedDate.set({ month: i });
-    months.push(formatter.format(date.toDate(state.timeZone)));
-  }
-
-  const selectedValue = {
-    month: months[state.focusedDate.month - 1].substring(0, 3),
-    year: state.focusedDate.year,
-  };
+  const ViewMap = {
+    month: <MonthDropdown setSelectedDropdown={setSelectedDropdown} />,
+    year: <YearDropdown setSelectedDropdown={setSelectedDropdown} />,
+  } satisfies { [key in ViewMapKeys]: React.JSX.Element };
 
   return (
-    <div
-      tabIndex={-1}
+    <Calendar
       className={cn(
         'flex min-h-[350px] w-[360px] flex-col rounded-sm p-4  shadow-[0_4px_4px_rgba(165,165,165,0.25)]',
         classNames.calendar
       )}
-      {...calendarProps}
-      {...calendarState}
-      ref={ref}
+      {...props}
     >
       {selectedDropdown ? (
-        selectedDropdown === 'month' ? (
-          <MonthDropdown
-            setSelectedDropdown={setSelectedDropdown}
-            months={months}
-            state={state}
-          />
-        ) : (
-          <YearDropdown
-            setSelectedDropdown={setSelectedDropdown}
-            state={state}
-          />
-        )
+        ViewMap[selectedDropdown]
       ) : (
         <>
-          <div className="mb-4 flex items-center justify-between">
+          <header className="mb-4 flex items-center justify-between">
             <div className="flex w-full gap-4">
-              <button
-                disabled={state.isDisabled}
-                onClick={() => setSelectedDropdown('month')}
-                className={cn(buttonStyles, selectClassNames)}
-                data-testid="month"
-              >
-                {selectedValue.month}
-                <ChevronDown />
-              </button>
-              <button
-                disabled={state.isDisabled}
-                onClick={() => setSelectedDropdown('year')}
-                className={cn(buttonStyles, selectClassNames)}
-                data-testid="year"
-              >
-                {selectedValue.year}
-                <ChevronDown />
-              </button>
+              <CalendarButtonListBox
+                type="month"
+                isDisabled={props.isDisabled}
+                setSelectedDropdown={setSelectedDropdown}
+              />
+              <CalendarButtonListBox
+                type="year"
+                isDisabled={props.isDisabled}
+                setSelectedDropdown={setSelectedDropdown}
+              />
             </div>
             <div
               className={cn(
@@ -151,17 +93,44 @@ export const Calendar = ({
                 classNames.calendarControllers
               )}
             >
-              <Button {...prevPropsRest} disabled={prevIsDisabled}>
+              <Button
+                className={cn(
+                  'inline-flex items-center justify-center gap-[0.5ch]',
+                  buttonClassNames
+                )}
+                slot="previous"
+              >
                 <ChevronLeft />
               </Button>
-              <Button {...nextPropsRest} disabled={nextIsDisabled}>
+              <Button
+                className={cn(
+                  'inline-flex items-center justify-center gap-[0.5ch]',
+                  buttonClassNames
+                )}
+                slot="next"
+              >
                 <ChevronRight />
               </Button>
             </div>
-          </div>
-          <CalendarGrid state={state} />
+          </header>
+          <CalendarGrid>
+            <CalendarGridHeader />
+            <CalendarGridBody>
+              {date => (
+                <CalendarCell
+                  date={date}
+                  className={cn(
+                    'flex h-[30px] w-[30px] cursor-pointer items-center justify-center rounded-full p-[5.3px] text-sm font-normal aria-disabled:cursor-default',
+                    classNames.calendarCell
+                  )}
+                />
+              )}
+            </CalendarGridBody>
+          </CalendarGrid>
         </>
       )}
-    </div>
+    </Calendar>
   );
 };
+
+export { _Calendar as Calendar };
