+++ packages/components/src/Calendar/index.ts
@@ -1,4 +1,1 @@
 export * from './Calendar';
-export * from './Calendar';
-export * from './CalendarCell';
-export * from './CalendarGrid';
